lolo

+++ yl-search-service/src/main/java/com/yoloo/server/search/queue/handler/DeletePostEventHandler.java
@@ -1,7 +1,7 @@
 package com.yoloo.server.search.queue.handler;
 
-import com.yoloo.server.common.queue.api.EventType;
-import com.yoloo.server.common.queue.api.YolooEvent;
+import com.yoloo.server.common.queue.vo.EventType;
+import com.yoloo.server.common.queue.vo.YolooEvent;
 import com.yoloo.server.search.post.PostRepository;
 
 import java.util.List;
@@ -11,24 +11,23 @@ public class DeletePostEventHandler extends EventHandler {
   private final PostRepository postRepository;
 
   public DeletePostEventHandler(EventType eventType, PostRepository postRepository) {
-    super(eventType);
+    super();
     this.postRepository = postRepository;
   }
 
   @Override
-  protected boolean isHandled(EventType eventType) {
-    return eventType == EventType.DELETE_POST;
-  }
+  public void process(EventType eventType, List<YolooEvent> events) {
+    if (eventType == EventType.DELETE_POST) {
+      List<String> postIds =
+          events
+              .stream()
+              .map(YolooEvent::getPayload)
+              .map(map -> (String) map.get("id"))
+              .collect(Collectors.toList());
 
-  @Override
-  public void process(List<YolooEvent> events) {
-    List<String> postIds =
-        events
-            .stream()
-            .map(YolooEvent::getPayload)
-            .map(map -> (String) map.get("id"))
-            .collect(Collectors.toList());
+      postRepository.deletePostsByIdIn(postIds);
+    }
 
-    postRepository.deletePostsByIdIn(postIds);
+    processNext(eventType, events);
   }
 }
